update fetch location
The ".." trick is useful when dealing with a manifest repo that might
bounce around server instances and relative sub-paths (like Android
where things might be mirrored under a /foo/ prefix everywhere).  In
the Weave world, we don't care about this as we only have one GoB
instance to hold all our code.

This also simplifies buildbot setup as the relative path causes it to
try and fetch from the wrong location.

Bug: 28346734

Change-Id: Id3ef1381d1f0efc18bca9a0ed45e6dafbf7f95e2
Reviewed-on: https://weave-review.googlesource.com/3415
Reviewed-by: Bhaskar Borthakur <borthakur@google.com>
Reviewed-by: Robert Ginda <rginda@google.com>